I ACTUALLY FIXED A BUG. IT WAS THE RESULT BEING AN INT INSTEAD OF A LONG

+++ src/exec/httpget.c
@@ -8,21 +8,21 @@
 #include "net.h"
 
 int main(int argc, char **argv) {
-  int res;
+  long res;
   if (argc != 3) {
     printf("%s url outfile", argv[0]);
     return(1);
   }
   res = net_init();
   if (res != 0) {
-    printf("net_init() error: %d\n", res);
+    printf("net_init() error: %ld\n", res);
     return(1);
   }
   res = http_get(argv[1], argv[2], NULL, 8080, "Downloading %s... %ld bytes");
   if (res >= 0) {
-    printf("Saved %d bytes into %s.\n", res, argv[2]);
+    printf("Saved %ld bytes into %s.\n", res, argv[2]);
   } else {
-    printf("ERROR OCCURED: %d\n", res);
+    printf("ERROR OCCURED: %ld\n", res);
   }
   return(0);
 }
