+++ src/http.h
@@ -6,6 +6,6 @@
 #ifndef HTTP_GET_H_SENTINEL
 #define HTTP_GET_H_SENTINEL
 
-int http_get(char *orgurl, char *outputfile, char *proxy, int proxyport, char *downloadingstring);
+long http_get(char *orgurl, char *outputfile, char *proxy, int proxyport, char *downloadingstring);
 
 #endif
